vista de salones implmentada, falta la vista detalle de salones

+++ js/funcionesClientes.js
@@ -2,23 +2,29 @@ const baseUrl = 'http://localhost:8080/api/Client/'
 
 window.onload = async function(){
 
-	const tableData = document.querySelector(".tabla_clientes")
+	const tableData = document.querySelector(".table_body")
 	const formClient = document.querySelector(".formulario_clientes");
 	var clientsEditOption = [];
 	var clientsDeleteOption = [];
+	var clientsViewOption = [];
+	var formOptionActive = false;
 
 	updateTable(tableData)
 	
 	/*Actualizar la tabla de registro*/
 	async function updateTable(table){
 		const data = await listarTodos(baseUrl);
-		const headers = table.querySelector(".table_headers");
-		table.innerHTML = "";
-		table.appendChild(headers);
-		for(var obj of await data){
-			table.innerHTML += formatRowTable(obj);
-		}
-		updateClientsOptions();
+		if(Symbol.iterator in Object(data) && data.length > 0){
+			document.querySelector(".without_results").style.display = "none";
+			table.innerHTML = "";
+			for(var obj of await data){
+				table.innerHTML += formatRowTable(obj);
+			}
+		}else{
+            table.innerHTML = ""
+            document.querySelector(".without_results").style.display = "block";
+        }
+        updateClientsOptions();
 	}
 
 	/*Crea un nuevo usuario*/
@@ -36,6 +42,8 @@ window.onload = async function(){
 		const response = await crearObjecto(baseUrl, data);
 
 		if(await response.idClient != null){
+			e.target.reset();
+			document.querySelector(".without_results").style.display = "none";
 			tableData.innerHTML += formatRowTable(await response);
 			updateClientsOptions();
 		}
@@ -46,16 +54,16 @@ window.onload = async function(){
 
 	/*Agrega los datos del usuario al formulario para poder ser modificados*/
 	async function updateClient(id){
+		formOptionActive = true;
 		const data = await buscarObjeto(baseUrl, id);
-		formClient.name_cliente.value = await data.name;
-		formClient.email_cliente.value = await data.email;
-		formClient.age_cliente.value = await data.age;
-		formClient.password_cliente.value = await data.password;
+		assignValuesInputs(await data.name, await data.email, await data.age, await data.password);
 		var submit_btn = formClient.querySelector(".submit_btn");
+		var cancel_btn = formClient.querySelector(".cancel_btn");
 		var title_form = formClient.querySelector(".form_title");
 		if(submit_btn){
 			submit_btn.innerHTML = "Actualizar";
 			submit_btn.classList.replace("submit_btn", "btn_disabled");
+			cancel_btn.style.display = "inline-block";
 		}
 		if(title_form){
 			title_form.innerHTML = "Actualizar Cliente"
@@ -67,6 +75,9 @@ window.onload = async function(){
 		/*Actualiza los datos de un usuario*/
 		formClient.removeEventListener("submit", createClient);
 		formClient.addEventListener("submit", updateClientAPI);
+
+		/*Cancela la actualizacion y resetea el form a su estado original*/
+		cancel_btn.addEventListener("click", resetForm)
 
 		async function updateClientAPI(e){
 			e.preventDefault();
@@ -84,20 +95,74 @@ window.onload = async function(){
 
 			if(await response.id !== null){
 				/*Restablece el formulario*/
-				title_form.innerHTML = "Registro de Clientes";
-				formClient.reset();
-				formClient.removeEventListener("submit", updateClientAPI)
-				formClient.addEventListener("submit", createClient);
-				updateTable(tableData)
+				resetForm();
+				updateTable(tableData);
+			}
+		}
+
+		/*Reseteo del formulario*/
+		function resetForm(){
+			title_form.innerHTML = "Registro de Clientes";
+			formClient.reset();
+			formClient.removeEventListener("submit", updateClientAPI)
+			formClient.addEventListener("submit", createClient);
+			cancel_btn.removeEventListener("click", resetForm);
+			cancel_btn.style.display = "none";
+			submit_btn.innerHTML = "Registrar";
+			formOptionActive = false;
+		}
+	}
+
+	/*Eliminar cliente*/
+	async function deleteClient(id){
+		if(confirm("Seguro que desea eliminar este cliente")){
+			const response = await eliminarObjecto(baseUrl, id);
+			if(await response === 204){
+				updateTable(tableData);
 			}
 		}
 	}
 
-	async function deleteClient(id){
-		const response = await eliminarObjecto(baseUrl, id);
-		if(response){
-			alert("cliente eliminado");
-			updateTable(tableData);
+	/*Vista detalle del cliente*/
+	async function viewDetails(id){
+		formOptionActive = true;
+		const data = await buscarObjeto(baseUrl, id);
+		if(data != null){
+			const submit_btn = formClient.querySelector(".submit_btn")
+			submit_btn.style.display = "none";
+
+			var title_form = formClient.querySelector(".form_title");
+			title_form !== null? title_form.innerHTML = "Cliente" : "";
+			
+			var cancel_btn = formClient.querySelector(".cancel_btn");
+			cancel_btn !== null? cancel_btn.innerHTML = "Cerrar" : "";
+			cancel_btn.style.display = "inline-block";
+
+			assignValuesInputs(data.name, data.email, data.age, data.password);
+			
+			const fieldForms = document.querySelectorAll(".field_form");
+			fieldForms.forEach(field => {
+				field.classList.add("field_form_view");
+				field.querySelector("input").classList.replace("input_form", "form_data_view");
+			})
+
+			cancel_btn.addEventListener("click", resetForm);
+
+			/*Reseteo del formulario*/
+			function resetForm(){
+				title_form.innerHTML = "Registro de Clientes";
+				assignValuesInputs("", "", "", "");
+				fieldForms.forEach(field => {
+					field.classList.remove("field_form_view");
+					field.querySelector("input").classList.replace("form_data_view", "input_form");
+				})
+				cancel_btn.removeEventListener("click", resetForm);
+				cancel_btn.style.display = "none";
+				cancel_btn.innerHTML = "Cancelar";
+				submit_btn.style.display = "inline-block";
+				formOptionActive = false;
+			}
+
 		}
 	}
 
@@ -112,6 +177,7 @@ window.onload = async function(){
 			<td class="columna-options-tabla-clientes">
 				<i class="far fa-edit vista-icon icon edit_client" data-id="${data.idClient}"></i>
 				<i class="fas fa-trash-alt eliminar_icon icon delete_client" data-id="${data.idClient}"></i>
+				<i class="far fa-window-maximize view_icon icon view_client" data-id="${data.idClient}"></i>
 			</td>
 		</tr>
 		`
@@ -121,162 +187,18 @@ window.onload = async function(){
 	function updateClientsOptions(){
 		clientsEditOption = document.querySelectorAll(".edit_client");
 		clientsDeleteOption = document.querySelectorAll(".delete_client");
-		clientsEditOption.forEach(btn => btn.addEventListener("click", () => updateClient(btn.getAttribute("data-id"))));
+		clientsViewOption = document.querySelectorAll(".view_client");
+		clientsEditOption.forEach(btn => btn.addEventListener("click", () => formOptionActive? "" : updateClient(btn.getAttribute("data-id"))));
 		clientsDeleteOption.forEach(btn => btn.addEventListener("click", () => deleteClient(btn.getAttribute("data-id"))));
+		clientsViewOption.forEach(btn => btn.addEventListener("click", () => formOptionActive? "" : viewDetails(btn.getAttribute("data-id"))));
+	}
+
+	/*asign values to form inputs*/
+	function assignValuesInputs(name, email, age, password){
+		formClient.name_cliente.value = name;
+		formClient.email_cliente.value = email;
+		formClient.age_cliente.value = age;
+		formClient.password_cliente.value = password;
 	}
 
 }
-
-// function listarClientes(){
-
-// 	$.ajax({
-// 	url: `${baseUrl}/ords/admin/client/client`,
-// 	data: '{}',
-//     type : 'GET',
-//     dataType : 'json',
-//     contentType: 'application/json; charset=utf-8',
-// 	success:function(respuesta){
-// 		let tabla;
-// 		let clientes = respuesta.items
-// 		for(i=0; i<clientes.length; i++){
-// 			tabla += '<tr>';              
-// 			tabla += '<td class="columna-id-tabla-clientes">'+respuesta.items[i].id+'</td>';
-// 			tabla += '<td class="columna-name-tabla-clientes">'+respuesta.items[i].name+'</td>';
-// 			tabla += '<td class="columna-email-tabla-clientes">'+respuesta.items[i].email+'</td>';
-// 			tabla += '<td class="columna-age-tabla-clientes">'+respuesta.items[i].age+'</td>';
-// 			tabla += `<td class="columna-options-tabla-clientes"><i class="fas fa-trash-alt eliminar-icon icon" onclick="eliminarCliente(${respuesta.items[i].id})"></i><i class="fas fa-envelope-open-text vista-icon icon" onclick="verCliente(${respuesta.items[i].id})"></i></td>`;
-// 			tabla +='</tr>';			
-// 		}
-// 		$(".tabla_clientes").append(tabla);
-// 	},
-// 	error: function(xhr,status){
-// 		console.log('error' + status);
-// 	}
-// 	}
-	
-// 	)
-// }
-
-// function agregarCliente(){
-// 	$.ajax({
-// 		url : `${baseUrl}/ords/admin/client/client`,
-// 		type: 'POST',
-// 		dataType: 'json',
-// 		contentType: 'application/json; charset=utf-8',
-// 		data: JSON.stringify({
-// 			id: parseInt($("#id_cliente").val()),
-// 			name: $("#name_cliente").val(),
-// 			email: $("#email_cliente").val(),
-// 			age: parseInt($("#age_cliente").val())
-// 		}),
-// 		statusCode: {
-// 			201: () => listarClientes()
-// 		}
-// 	})
-// }
-
-// // AL ENVIAR EL FORMULARIO DE REGISTRO
-// $(".formulario-clientes").submit(function(e){
-// 	e.preventDefault();
-// 	agregarCliente();
-// })
-
-// $(".registrar-clientes-boton").click(function(e){
-// 	e.preventDefault();
-// 	agregarCliente();
-// })
-
-// // ELIMINACION DE CLIENTES
-// function eliminarCliente(id){
-// 	$.ajax({
-// 	  url: `${baseUrl}/ords/admin/client/client`,
-// 	  type: 'DELETE',
-// 	  data: JSON.stringify({ id: id }),
-// 	  dataType: 'json',
-// 	  contentType: 'application/json; charset=utf-8',
-// 	  statusCode: {
-// 		204: () => location.reload(true)
-// 	  },
-// 	})
-// }
-
-// // VISTA DETALLE DE CLIENTES
-// function verCliente(id){
-// 	$.ajax({
-// 		url: `${baseUrl}/ords/admin/client/client`,
-// 		data: '{}',
-// 		type : 'GET',
-// 		dataType : 'json',
-// 		contentType: 'application/json; charset=utf-8',
-// 		success:function(respuesta){
-// 			let cliente;
-// 			let clientes = respuesta.items;
-// 			for(i=0; i<clientes.length; i++){
-// 				if(respuesta.items[i].id == id) {
-// 					cliente = respuesta.items[i];
-// 					$(".container-data").empty();
-// 					$(".container-data").append(`
-// 						<h1>Cliente</h1>
-// 						<i id="update-data" class="fas fa-user-cog update-data" onclick="vistaActualizarCliente()"></i>
-// 						<label for="id_cliente" class="label-clientes label-data-cliente">Id: <span id="cliente_id" class="user-data"> ${cliente.id}</span></label>
-// 						<label for="name_cliente" class="label-clientes label-data-cliente">Nombre: <span id="cliente_name" class="user-data"> ${cliente.name}</span></label>
-// 						<label for="email_cliente" class="label-clientes label-data-cliente">Email:  <span id="cliente_email" class="user-data"> ${cliente.email}</span></label>
-// 						<label for="age_cliente" class="label-clientes label-data-cliente">Edad:  <span id="cliente_age" class="user-data"> ${cliente.age}</span></label>
-// 						<button class="cerrar-data-boton" onclick="cerrarData()">Cerrar</button>
-// 					`)
-// 					break;
-// 				}		
-// 			}
-// 		},
-// 		error: function(xhr,status){
-// 			console.log('error' + status);
-// 		}
-// 		})
-// }
-
-// // CERRAR PANEL DE VISTA DETALLE DE CLIENTES
-// function cerrarData(){
-// 	location.reload(true);
-// }
-
-// // PASAR DE VISTA DETALLE A VISTA ACTUALIZACION
-// function vistaActualizarCliente(){
-// 	let cliente =  {id:$("#cliente_id").text(), name: $("#cliente_name").text(), email: $("#cliente_email").text(), age: parseInt($("#cliente_age").text())}
-// 	$(".container-data").empty();
-// 	$(".container-data").append(`
-// 		<h1>Actualizar Cliente</h1>
-// 		<label for="id_cliente" class="label-clientes label-data-cliente">Id: 
-// 			<span id="cliente_id" class="user-data"> ${cliente.id}</span>
-// 		</label>
-// 		<label for="name_cliente" class="label-clientes label-data-cliente">Nombre: 
-// 			<input type="text" id="name_cliente" name="name_cliente" class="input-clientes" value="${cliente.name}" required/>
-// 		</label>
-// 		<label for="email_cliente" class="label-clientes label-data-cliente">Email:  
-// 			<input type="email" id="email_cliente" name="email_cliente" class="input-clientes" value="${cliente.email}" required/>
-// 		</label>
-// 		<label for="age_cliente" class="label-clientes label-data-cliente">Edad:  
-// 			<input type="number" id="age_cliente" name="age_cliente" class="input-clientes" value="${cliente.age}" required/>
-// 		</label>
-// 		<button class="actualizar-data-boton" onclick="actualizarCliente()">Actualizar</button>
-// 	`)
-// }
-  
-
-// function actualizarCliente(){
-// 	$.ajax({
-// 		url : `${baseUrl}/ords/admin/client/client`,
-// 		type: 'PUT',
-// 		dataType: 'json',
-// 		contentType: 'application/json; charset=utf-8',
-// 		data: JSON.stringify({
-// 			id: parseInt($("#cliente_id").text()),
-// 			name: $("#name_cliente").val(),
-// 			email: $("#email_cliente").val(),
-// 			age: parseInt($("#age_cliente").val())
-// 		}),
-// 		statusCode: {
-// 			201: () => {listarClientes();verCliente(parseInt($("#cliente_id").text()))}
-// 		}
-// 	})
-// }
-
